refactoring HTTP protocol
